Bug #680225 Coverity sensor otp warning resolution, manual clone from gerrit#381325
[root cause  ] code warnings
[changes     ] fix logic
[side effects]NA
[self test   ]<board, Test cases summary>:ok
[reviewers   ] miller.somg
[change_type ] feature_bugfix
[tag_product ] common

+++ sensor/otp_drv/otp_common.c
@@ -59,7 +59,7 @@ static cmr_u8 *otp_formatted_data_buffer[4] = {NULL, NULL, NULL, NULL};
  **/
 
 cmr_int sensor_otp_rw_data_from_file(cmr_u8 cmd, char *file_name,
-                                     void **otp_data, int *otp_size) {
+                                     void **otp_data, long *otp_size) {
     cmr_int ret = OTP_CAMERA_SUCCESS;
     OTP_LOGI("in");
     CHECK_PTR((void *)otp_data);
